#1 restructing the form for progres report
