+++ tajma/forms/PsychometricTests/ProfileForm.py
@@ -1,0 +1,8 @@
+from flask_wtf import FlaskForm 
+from flask_wtf.file import FileAllowed
+from wtforms import FileField, SubmitField
+
+class UpdateAccountForm(FlaskForm):
+    picture = FileField('Update Profile Picture', validators=[FileAllowed(['jpg', 'png'])])
+    #email = StringField('Email', validators=[DataRequired(), Email()])
+    update = SubmitField('Update')
