+++ tajma/forms/PsychometricTests/__init__.py
@@ -1,0 +1,18 @@
+import imp
+from flask_wtf import FlaskForm
+
+class FieldsRequiredForm(FlaskForm):
+    """Require all fields to have content. This works around the bug that WTForms radio
+    fields don't honor the `DataRequired` or `InputRequired` validators.
+    """
+
+    class Meta:
+        def render_field(self, field, render_kw):
+            render_kw.setdefault('required', True)
+            return super().render_field(field, render_kw)
+
+
+from .AttitudeForm import AttitudeAnswer
+from .ElearningForm import ElearningAnswer
+from .LearnerForm import LearnerAnswer
+from .ProfileForm import UpdateAccountForm
