--- tajma/services/__init__.py
@@ -1,2 +1,0 @@
-def init_app(app):
-    return
