--- tajma/templates/SaveProgress.html
@@ -1,8 +1,0 @@
-
-{%extends 'dashboard.html'%}
-{%block body%}
-<div class="container-fluid">
-    <h1>Progress Saved</h1>
-</div>
-
-{%endblock%}
