+++ tajma/templates/tp2Attitude.html
@@ -58,7 +58,7 @@
                     <li>The first answer that come through your mind is the most accurate</li>
                     <li>Good luck</li>
                     <button id="btn-close-modal" type="button" class="close" data-dismiss="modal" aria-label="Close">Proceed to take Second test</button>
-                    <a type="button" class="close" data-dismiss="modal" aria-label="Close" href="{{url_for('dashboard.dashboard.query', test="elearner")}}">Save First Test</button>
+                    <a type="button" class="close" data-dismiss="modal" aria-label="Close" href="{{url_for('dashboard.dashboardWithFlash', test="E-Learner")}}">Save First Test</button>
                 </ul>
             </div>
         </div>
